code review comments resolved
